Changed placeholder double type for "Shooter" to AnalogChannel and used correct library functions; working build.
We're amazed it worked.

+++ CodeRedRobot2012/Subsystems/Shooter.cpp
@@ -10,7 +10,7 @@ void Shooter::InitDefaultCommand() {
 	//SetDefaultCommand(new MySpecialCommand());
 }
 
-void Shooter::Shoot(double power) {
+void Shooter::Shoot(float power) {
 	
 }
 // Put methods for controlling this subsystem
